+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FEATURES: 
 
+* **New Data Source:** `azurerm_portal_dashboard` [GH-15326]
 * **New Beta Resource:** `azurerm_app_service_source_control_slot` [GH-15301]
 * **New Resource:** `azurerm_sentinel_watchlist_item` [GH-14366]
 * **New Resource:** `azurerm_stream_analytics_output_function` [GH-15162]
@@ -10,13 +11,16 @@ FEATURES:
 ENHANCEMENTS: 
 
 * dependencies: updating to `v0.23.1` of `github.com/hashicorp/go-azure-helpers` [GH-15314]
+* `azurerm_application_gateway` - the field `type` within the `identity` block is now required when an `identity` block is specified [GH-15337]
 * `azurerm_application_insights` - Add support for `force_customer_storage_for_profiler` [GH-15254]
 * `azurerm_automation_account` - Support for managed identity [GH-15072]
+
 
 BUG FIXES:
 
 * `azurerm_application_insights_api_key` - prevent panic by checking for the id of an existing API Key [GH-15297]
 * `azurerm_app_service_active_slot` - fix regression in ID set in creation of new resource [GH-15291]
+* `azurerm_firewall` - working around an Azure API issue when deleting the Firewall [GH-15330]
 * `azurerm_orchestrated_virtual_machine_scale_set` - fixing a crash when the `computer_name_prefix` wasn't specified [GH-15312]
 * `azurerm_recovery_services_vault` - fixing an issue where the subscription couldn't be found when running in Azure Government [GH-15316]